backend/requirements.txt
# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.1.0

# External APIs
anthropic>=0.18.0
gradio-client>=0.10.0
google-api-python-client>=2.114.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Object Storage
minio>=7.2.3

# Task Queue
redis>=5.0.1
rq>=1.15.1

# Scheduling
apscheduler>=3.10.4

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
tenacity>=8.2.3
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0
factory-boy>=3.3.0
faker>=22.0.0

# Development
black>=24.1.0
isort>=5.13.2
mypy>=1.8.0
ruff>=0.1.14
